+++ settings.js
@@ -63,6 +63,8 @@ var Options = class Options {
             enablePageShortcuts: ['boolean', 'enable-page-shortcuts'],
             showWsSwitcherBg: ['boolean', 'show-ws-switcher-bg'],
             showWsPreviewBg: ['boolean', 'show-ws-preview-bg'],
+            showBgInOverview: ['boolean', 'show-bg-in-overview'],
+            overviewBgBlurSigma: ['int', 'overview-bg-blur-sigma'],
             appGridAnimation: ['int', 'app-grid-animation'],
             workspaceAnimation: ['int', 'workspace-animation'],
             fixUbuntuDock: ['boolean', 'fix-ubuntu-dock'],